muutamia tarpeettomia juttuja poistettu

+++ app/src/main/java/com/example/fiilismittari/DataPoint.java
@@ -1,15 +1,15 @@
 package com.example.fiilismittari;
 //datapoint will have the mood and date
 public class DataPoint {
-    private int mood;
+    private double mood;
     private String date;
 
-    public DataPoint(int mood, String date) {
+    public DataPoint(double mood, String date) {
         this.mood = mood;
         this.date = date;
     }
 
-    public int getMood() {
+    public double getMood() {
         return mood;
     }
 }
